Enable Redirect on checkout
Razorpay doesnot redirect the user to the plugin endpoint in case of
filed payments by default.

Adding a limit of 2 failed payments that after which the user will be
redirected to the checkout page.

Resoning: A lot of payments were ending up in "Awaiting Payment" stage
because failed payments were not being captured. This prevented the
users from visiting the registration link and retry the payment, and
required registration managers to manually mark the payment as "Not
Paid".

This will prevent some of such events.